Updated and fixed an issue
The issue should be fixed once the server will be using Easy RPG Player 0.72 or the latest development build. Until that, the translation will work fine, but the bug will still be present. The issue was caused by "Actor's name is (name)" not being translated until the latest builds of Easy RPG.