[Yume 2kki - Spanish] Translation update (#48)
* Translated GB pictures in Spanish using actual translation

There's just the Change Screen Color that I did not entirely translated since I don't speak Spanish.

* Added every map Effects and Menu translations in Spanish

* Added almost all files of Jigsaw Puzzle World

Every effects are translated in Spanish, and 4 on 8 misc. pictures are translated.

* Added ¿ at the start of interrogative sentences

* Updated file and fixed an issue

Fixed an issue which could cause a softlock in a specific occasion
The issue should be fixed once the server will be using Easy RPG Player 0.72 or the latest development build. Until that, the translation will work fine, but the bug will still be present. The issue was caused by "Actor's name is (name)" not being translated until the latest builds of Easy RPG.

* Updated Sound Room

Still use English translation, just updated the file.

* Finalized Jigsaw Puzzle and Monochrome GB World

* Finalized text

- Fixed the error in "Tema de Menú", mispelled as "Temas de Menú"
- Translated key item
- Translated Menu Themes of Maple Shrine and Blob Desert
- Finalized translation of the Quick effect setting.

* Removed space to the Key item since it does not normally exist

* Added Quick Effect Menu

* Fixed a text display issue