Added FC Menu assets
Everything being based on old assets, this is a placeholder.
The Dice effect was added in grey, but it is not present on the original menu (last Nuaah update was two months after the effect was added, hence why it is not present on the original asset).
The Russian translation will also likely need to be updated since the Dice effect is not there.
This PR only covers the English, French, Spanish, Romanian and Brazilian Portuguese translations, Russian was already made by Fonvight, and I don't know how to draw Korean and Chinese characters.
Finally, I've not translated the error: I'm not sure how to translate it, and since it's not a part of the UI it's not a big issue if it's left untranslated.