Added the English translation of R.B.Y. to the French translation
Most of the games from the NES period without a major plot that needs to be understand tend to be only available in English in France (or only with the manual translated), hence why I've not translated the assets directly in French here. I don't see it being translated in French.