Initial Commit
Should be complete, but is once again ultimately untested.
A note: Famicom-like areas use Hiragana to spell out most things, as if working under a character limitation where Kanji or Katakana character sets are unavailable. Localizing this effect was given low priority, as an equivalent for the effect in Spanish didn't occur to me.