Mikan Muzou Spanish Translation (#158)
* Initial Commit

Should be complete, but is once again ultimately untested.
A note: Famicom-like areas use Hiragana to spell out most things, as if working under a character limitation where Kanji or Katakana character sets are unavailable. Localizing this effect was given low priority, as an equivalent for the effect in Spanish didn't occur to me.

* Uncentering Quit, making FC Quit msg match

What it says on the tin: I don't feel these were issues, personally.
* Title screen's quit is no longer centered via two spaces.
* FC's quit screen's kerning has been made to match most the rest of the area's (sans Distortion), and the content's been made to match the game's normal quit screen's.

* Drop Shadowing No

Whoops. Fixed.