If - Add a French translation (#240)
* If (English) - Fix the first instructions using the wrong text for waking up

Pinch Cheeks is only used for the second part of the game where the other set of instructions is used, Twist Cheeks is for the first part with those instructions.

* If (English) - Remove unused files

Map 60 and map 69 contain lines only used in debug but aren't translated, so they aren't needed.
RPG_RT.ldb.common does not contain anything, RPG_RT.ldb.battle only contain a line for displaying the money which doesn't need to be edited, and RPG_RT.lmt contain the names of the maps which isn't used, so all of those files aren't needed.

* If - Add a French translation