Yume Nikki - Fix the Hebrew translation (for now) (#246)
Fix the display of the Hebrew translation, by adding a font, and rewriting how the text is displayed. Change the name of the language on the menu to Hebrew due to not being able to be displayed until being selected if the language was written in its original language, and remove useless files. Adjust the spacing of some lines to be centered.
While this will work for some time, due to a lack of Bidi-algorithm in the EasyRPG Player at the moment (https://github.com/EasyRPG/Player/issues/2982), every description in the RPG_RT.ldb.po file (+ save, load, quit lines) will have to be edited another time due to the . character altering how the text is written + how descriptions are handled, so I had to edit lines to take into account this.
Fix to do for the future once said algorithm is supported: take the RPG_RT.ldb.po version of this translation prior to this pull request, take all of the lines of text, and revert the text without having to deal with manually adjusting the text.