Add translated lines outside of the common file
Map 0001: For some reason, the line

Please choose the option that you prefer to see for collecting crystals.

added a text to say that you need to press the cancel key to cancel after, and in the choice after,

Text only
Animation only
Text and animation
Cancel

the cancel option was removed. Was undone here by removing the line added and using the cancel line from somewhere else