Someday - update the Korean translation to 0.11 (#250)
* Someday - Update the graphical assets of the Korean translation

Ignored:
- alternative text asset for the Clockworker effect (seems to be wip version of the one used and not used in-game)
- Error graphic (doesn't seem to be used either, so perhaps something made by Nuls for fun)

* Add translated lines outside of the common file

Map 0001: For some reason, the line

Please choose the option that you prefer to see for collecting crystals.

added a text to say that you need to press the cancel key to cancel after, and in the choice after,

Text only
Animation only
Text and animation
Cancel

the cancel option was removed. Was undone here by removing the line added and using the cancel line from somewhere else

* - Add the fixed pictures from the revision A

* Update all of the remaining text